refactor(request-validator): split into per-transport domain classes
Step 2 of 3 for #68 (the prior PR #90 landed the shared Validation\Support
helpers). Per-transport validation logic is now isolated in six dedicated
classes under `Studio\OpenApiContractTesting\Validation\Request\`:

- ParameterCollector       — path+operation merge, malformed-entry flagging,
                             OAS §4.7.12.1 reserved in:header rejection
- PathParameterValidator   — in:path validation (inc. rawurldecode + placeholder
                             declaration cross-check)
- QueryParameterValidator  — in:query validation (form+explode:true only)
- HeaderParameterValidator — in:header validation (case-insensitive match,
                             HeaderBag unwrap, multi-value/non-scalar guards)
- RequestBodyValidator     — requestBody + content-type negotiation + JSON
                             schema check
- SecurityValidator        — operation/root `security` resolution, OR/AND
                             semantics, http+bearer / apiKey(header|query|
                             cookie) satisfaction; oauth2/OIDC/mTLS are
                             skipped as unsupported; classifyScheme /
                             checkSchemeSatisfaction / extractSingleStringValue
                             stay private to the class.

OpenApiRequestValidator shrinks from 1034 → 121 lines and is now a pure
orchestrator: load spec → match path → check method → collect parameters →
delegate to the five validators → concatenate errors → return a single
OpenApiValidationResult. All five new validators take SchemaValidatorRunner
via constructor injection (SecurityValidator has no opis dependency).

Public API — `new OpenApiRequestValidator(int $maxErrors = 20)` and
`validate(...)` — is unchanged, so callers (including the Laravel
`ValidatesOpenApiSchema` trait, which only uses the response validator but
is covered by the same integration suite) need no modifications.

Tests:
- existing OpenApiRequestValidatorTest (153 tests, 322 assertions) is
  untouched and still passes — behavior preserved.
- six new focused unit tests under tests/Unit/Validation/Request/
  demonstrate that each extracted class is independently testable without
  orchestrator setup (23 new tests).

Full suite: 559 tests / 1165 assertions pass, phpstan level 6 clean,
php-cs-fixer clean.

Refs #68
