+++ src/OpenApiRequestValidator.php
@@ -4,48 +4,36 @@ declare(strict_types=1);
 
 namespace Studio\OpenApiContractTesting;
 
-use LogicException;
-use Studio\OpenApiContractTesting\Validation\Support\ContentTypeMatcher;
-use Studio\OpenApiContractTesting\Validation\Support\HeaderNormalizer;
-use Studio\OpenApiContractTesting\Validation\Support\ObjectConverter;
+use Studio\OpenApiContractTesting\Validation\Request\HeaderParameterValidator;
+use Studio\OpenApiContractTesting\Validation\Request\ParameterCollector;
+use Studio\OpenApiContractTesting\Validation\Request\PathParameterValidator;
+use Studio\OpenApiContractTesting\Validation\Request\QueryParameterValidator;
+use Studio\OpenApiContractTesting\Validation\Request\RequestBodyValidator;
+use Studio\OpenApiContractTesting\Validation\Request\SecurityValidator;
 use Studio\OpenApiContractTesting\Validation\Support\SchemaValidatorRunner;
-use Studio\OpenApiContractTesting\Validation\Support\TypeCoercer;
 
-use function array_key_exists;
-use function array_key_first;
 use function array_keys;
-use function array_values;
-use function count;
-use function get_debug_type;
-use function implode;
-use function in_array;
-use function is_array;
-use function is_int;
-use function is_scalar;
-use function is_string;
-use function preg_match;
-use function rawurldecode;
-use function sprintf;
 use function strtolower;
 
 final class OpenApiRequestValidator
 {
-    /**
-     * Per OpenAPI 3.x: "If `in` is `header` and the `name` field is `Accept`,
-     * `Content-Type` or `Authorization`, the parameter definition SHALL be
-     * ignored." These are controlled by content negotiation and security
-     * schemes; surfacing them through parameter validation would duplicate
-     * (and often disagree with) the framework's own handling.
-     */
-    private const RESERVED_HEADER_NAMES = ['accept', 'content-type', 'authorization'];
-
     /** @var array<string, OpenApiPathMatcher> */
     private array $pathMatchers = [];
-    private readonly SchemaValidatorRunner $runner;
+    private readonly PathParameterValidator $pathValidator;
+    private readonly QueryParameterValidator $queryValidator;
+    private readonly HeaderParameterValidator $headerValidator;
+    private readonly SecurityValidator $securityValidator;
+    private readonly RequestBodyValidator $bodyValidator;
 
     public function __construct(int $maxErrors = 20)
     {
-        $this->runner = new SchemaValidatorRunner($maxErrors);
+        $runner = new SchemaValidatorRunner($maxErrors);
+
+        $this->pathValidator = new PathParameterValidator($runner);
+        $this->queryValidator = new QueryParameterValidator($runner);
+        $this->headerValidator = new HeaderParameterValidator($runner);
+        $this->securityValidator = new SecurityValidator();
+        $this->bodyValidator = new RequestBodyValidator($runner);
     }
 
     /**
@@ -102,52 +90,19 @@ final class OpenApiRequestValidator
         /** @var array<string, mixed> $operation */
         $operation = $pathSpec[$lowerMethod];
 
-        // Collect merged path/operation parameters once so path + query validation
-        // share a single view of the spec and malformed-entry errors are surfaced
-        // only once.
-        [$parameters, $specErrors] = $this->collectParameters($method, $matchedPath, $pathSpec, $operation);
+        // Collect merged path/operation parameters once so path + query + header
+        // validation share a single view of the spec and malformed-entry errors
+        // are surfaced only once.
+        [$parameters, $specErrors] = ParameterCollector::collect($method, $matchedPath, $pathSpec, $operation);
 
-        $pathErrors = $this->validatePathParameters(
-            $method,
-            $matchedPath,
-            $parameters,
-            $pathVariables,
-            $version,
-        );
-        $queryErrors = $this->validateQueryParameters(
-            $method,
-            $matchedPath,
-            $parameters,
-            $queryParams,
-            $version,
-        );
-        $headerErrors = $this->validateHeaderParameters(
-            $method,
-            $matchedPath,
-            $parameters,
-            $headers,
-            $version,
-        );
-        $securityErrors = $this->validateSecurity(
-            $method,
-            $matchedPath,
-            $spec,
-            $operation,
-            $headers,
-            $queryParams,
-            $cookies,
-        );
-        $bodyErrors = $this->validateRequestBody(
-            $specName,
-            $method,
-            $matchedPath,
-            $operation,
-            $requestBody,
-            $contentType,
-            $version,
-        );
-
-        $errors = [...$specErrors, ...$pathErrors, ...$queryErrors, ...$headerErrors, ...$securityErrors, ...$bodyErrors];
+        $errors = [
+            ...$specErrors,
+            ...$this->pathValidator->validate($method, $matchedPath, $parameters, $pathVariables, $version),
+            ...$this->queryValidator->validate($method, $matchedPath, $parameters, $queryParams, $version),
+            ...$this->headerValidator->validate($method, $matchedPath, $parameters, $headers, $version),
+            ...$this->securityValidator->validate($method, $matchedPath, $spec, $operation, $headers, $queryParams, $cookies),
+            ...$this->bodyValidator->validate($specName, $method, $matchedPath, $operation, $requestBody, $contentType, $version),
+        ];
 
         if ($errors === []) {
             return OpenApiValidationResult::success($matchedPath);
@@ -162,873 +117,5 @@ final class OpenApiRequestValidator
     private function getPathMatcher(string $specName, array $specPaths): OpenApiPathMatcher
     {
         return $this->pathMatchers[$specName] ??= new OpenApiPathMatcher($specPaths, OpenApiSpecLoader::getStripPrefixes());
-    }
-
-    /**
-     * Validate query parameters declared by the matched operation (or
-     * inherited from the path-level `parameters` block).
-     *
-     * Only `style: form` + `explode: true` (the OpenAPI default for `in: query`)
-     * is supported. Repeated keys (`?tags=a&tags=b`) are expected to arrive as
-     * PHP arrays from the framework. Other styles (`form`+`explode:false`,
-     * `pipeDelimited`, `spaceDelimited`) are out of scope.
-     *
-     * @param list<array<string, mixed>> $parameters pre-collected merged parameters (path + operation level)
-     * @param array<string, mixed> $queryParams
-     *
-     * @return string[]
-     */
-    private function validateQueryParameters(
-        string $method,
-        string $matchedPath,
-        array $parameters,
-        array $queryParams,
-        OpenApiVersion $version,
-    ): array {
-        $errors = [];
-
-        foreach ($parameters as $param) {
-            if (($param['in'] ?? null) !== 'query') {
-                continue;
-            }
-
-            /** @var string $name */
-            $name = $param['name'];
-            $required = ($param['required'] ?? false) === true;
-
-            // A required parameter with no schema is a clearly malformed spec — surface it
-            // rather than silently passing every request. Optional parameters with no schema
-            // have nothing to validate, so we let them through (matches the body validator).
-            if (!isset($param['schema']) || !is_array($param['schema'])) {
-                if ($required) {
-                    $errors[] = "[query.{$name}] required parameter has no schema for {$method} {$matchedPath} — cannot validate.";
-                }
-
-                continue;
-            }
-
-            /** @var array<string, mixed> $schema */
-            $schema = $param['schema'];
-
-            $present = array_key_exists($name, $queryParams) && $queryParams[$name] !== null;
-            if (!$present) {
-                if ($required) {
-                    $errors[] = "[query.{$name}] required query parameter is missing.";
-                }
-
-                continue;
-            }
-
-            $coerced = TypeCoercer::coerceQuery($queryParams[$name], $schema);
-            $jsonSchema = OpenApiSchemaConverter::convert($schema, $version, SchemaContext::Request);
-
-            $schemaObject = ObjectConverter::convert($jsonSchema);
-            $dataObject = ObjectConverter::convert($coerced);
-
-            $formatted = $this->runner->validate($schemaObject, $dataObject);
-            foreach ($formatted as $path => $messages) {
-                $suffix = $path === '/' ? '' : $path;
-                foreach ($messages as $message) {
-                    $errors[] = "[query.{$name}{$suffix}] {$message}";
-                }
-            }
-        }
-
-        return $errors;
-    }
-
-    /**
-     * Validate path parameters declared by the matched operation (or
-     * inherited from the path-level `parameters` block) against the values
-     * extracted by the path matcher.
-     *
-     * Path parameters are always required in OpenAPI, so a declared `in: path`
-     * entry without a `schema` is treated as a hard spec error rather than a
-     * silent pass. Percent-encoded segments are decoded via `rawurldecode()`
-     * before type coercion / schema validation — the matcher leaves values
-     * raw so encoding policy stays in one place.
-     *
-     * String-valued `format: uuid | date-time | date | email | ...` is
-     * delegated to opis/json-schema's built-in FormatResolver (registered on
-     * the `string` type by default). Numeric OAS formats (`int32`, `int64`,
-     * `float`, `double`) are advisory-only and are not validated.
-     *
-     * @param list<array<string, mixed>> $parameters pre-collected merged parameters
-     * @param array<string, string> $pathVariables values extracted by OpenApiPathMatcher
-     *
-     * @return string[]
-     */
-    private function validatePathParameters(
-        string $method,
-        string $matchedPath,
-        array $parameters,
-        array $pathVariables,
-        OpenApiVersion $version,
-    ): array {
-        $errors = [];
-        $declared = [];
-
-        foreach ($parameters as $param) {
-            if (($param['in'] ?? null) !== 'path') {
-                continue;
-            }
-
-            /** @var string $name */
-            $name = $param['name'];
-            $declared[$name] = true;
-
-            // Defensive: every path placeholder in the matched template should have
-            // been captured by the regex. A mismatch here means the spec template and
-            // the compiled matcher disagree — surface it loudly rather than skipping.
-            if (!array_key_exists($name, $pathVariables)) {
-                $errors[] = "[path.{$name}] declared in {$method} {$matchedPath} spec but not captured by path matcher.";
-
-                continue;
-            }
-
-            if (!isset($param['schema']) || !is_array($param['schema'])) {
-                // Path parameters are implicitly required (OpenAPI spec), so a schema-less
-                // entry means every value passes — exactly the silent-drift outcome this
-                // library exists to prevent.
-                $errors[] = "[path.{$name}] parameter has no schema for {$method} {$matchedPath} — cannot validate.";
-
-                continue;
-            }
-
-            /** @var array<string, mixed> $schema */
-            $schema = $param['schema'];
-
-            $decoded = rawurldecode($pathVariables[$name]);
-            $coerced = TypeCoercer::coercePrimitive($decoded, $schema);
-            $jsonSchema = OpenApiSchemaConverter::convert($schema, $version, SchemaContext::Request);
-
-            $schemaObject = ObjectConverter::convert($jsonSchema);
-            $dataObject = ObjectConverter::convert($coerced);
-
-            $formatted = $this->runner->validate($schemaObject, $dataObject);
-            foreach ($formatted as $path => $messages) {
-                $suffix = $path === '/' ? '' : $path;
-                foreach ($messages as $message) {
-                    $errors[] = "[path.{$name}{$suffix}] {$message}";
-                }
-            }
-        }
-
-        // Reverse check: every `{placeholder}` in the URL template MUST be declared
-        // as an `in: path` parameter per OpenAPI. A captured-but-not-declared name
-        // means the spec author forgot the declaration entirely, which would otherwise
-        // let any value pass silently — the drift this library exists to catch.
-        foreach ($pathVariables as $name => $_) {
-            if (!isset($declared[$name])) {
-                $errors[] = "[path.{$name}] placeholder in {$method} {$matchedPath} template is not declared as an 'in: path' parameter — malformed spec (OpenAPI requires every placeholder to be declared).";
-            }
-        }
-
-        return $errors;
-    }
-
-    /**
-     * Validate header parameters declared by the matched operation (or
-     * inherited from the path-level `parameters` block).
-     *
-     * HTTP header names are case-insensitive (RFC 7230) so both the spec
-     * `name` and the caller-supplied `$headers` keys are lower-cased before
-     * matching. Error messages keep the spec's original casing so users can
-     * grep the spec directly.
-     *
-     * Per OpenAPI 3.x, `Accept`, `Content-Type`, and `Authorization`
-     * declarations are ignored — these are controlled by content negotiation
-     * and security schemes, not arbitrary header parameters.
-     *
-     * Header values arriving as `array<string>` (Laravel's HeaderBag models
-     * repeated occurrences this way) are unwrapped to a single value when the
-     * array holds exactly one element. Multi-value arrays against scalar
-     * schemas produce a hard error — real frameworks disagree on which of
-     * the repeated values "wins" (Laravel picks first, Symfony picks last),
-     * so silently picking one would mask a drift the contract test exists to
-     * expose. Empty arrays are treated as missing. `style: simple` with
-     * `type: array | object` is out of scope.
-     *
-     * @param list<array<string, mixed>> $parameters pre-collected merged parameters
-     * @param array<string, mixed> $headers caller-supplied request headers
-     *
-     * @return string[]
-     */
-    private function validateHeaderParameters(
-        string $method,
-        string $matchedPath,
-        array $parameters,
-        array $headers,
-        OpenApiVersion $version,
-    ): array {
-        $errors = [];
-        $normalizedHeaders = HeaderNormalizer::normalize($headers);
-
-        foreach ($parameters as $param) {
-            if (($param['in'] ?? null) !== 'header') {
-                continue;
-            }
-
-            /** @var string $name */
-            $name = $param['name'];
-            $lowerName = strtolower($name);
-
-            $required = ($param['required'] ?? false) === true;
-
-            // Same reasoning as query/path: a required parameter without a schema would
-            // silently pass every request, so surface it as a hard spec error. Optional
-            // entries without a schema have nothing to validate — let them through.
-            if (!isset($param['schema']) || !is_array($param['schema'])) {
-                if ($required) {
-                    $errors[] = "[header.{$name}] required parameter has no schema for {$method} {$matchedPath} — cannot validate.";
-                }
-
-                continue;
-            }
-
-            /** @var array<string, mixed> $schema */
-            $schema = $param['schema'];
-
-            $rawValue = $normalizedHeaders[$lowerName] ?? null;
-
-            // `null` and `[]` (empty repeated-header array) both collapse to "missing".
-            // A repeated header that was sent zero times is semantically absent.
-            if ($rawValue === null || $rawValue === []) {
-                if ($required) {
-                    $errors[] = "[header.{$name}] required header is missing.";
-                }
-
-                continue;
-            }
-
-            if (is_array($rawValue)) {
-                // HeaderBag shape: list<string>. Single-element arrays are the common
-                // case (Laravel always wraps) — unwrap. Multi-element means the client
-                // sent the header more than once; frameworks disagree on which value
-                // is "canonical" (Laravel: first, Symfony: last), so silently picking
-                // one would mask drift. Surface it so the spec author / client fixes
-                // the duplicate.
-                if (count($rawValue) > 1) {
-                    $errors[] = sprintf(
-                        '[header.%s] multiple values received (count=%d) but schema expects a single value; refusing to pick one silently.',
-                        $name,
-                        count($rawValue),
-                    );
-
-                    continue;
-                }
-
-                $rawValue = $rawValue[array_key_first($rawValue)];
-            }
-
-            // Mirror the pre-unwrap missing-header branch for the post-unwrap case:
-            // `['X-Foo' => [null]]` is a caller bug shaped identically to an absent
-            // header. Letting it flow to coercion would either silently pass against
-            // a `nullable` schema or surface as a `/` type mismatch from opis — both
-            // hide the root cause.
-            if ($rawValue === null) {
-                if ($required) {
-                    $errors[] = "[header.{$name}] required header is missing.";
-                }
-
-                continue;
-            }
-
-            // Guard against caller-side bugs that smuggle a non-scalar (nested array,
-            // object, resource) past the unwrap. Without this, opis would report a
-            // JSON-Pointer type mismatch that hides the real cause — that the caller
-            // never produced a header-shaped value in the first place.
-            if (!is_scalar($rawValue)) {
-                $errors[] = sprintf(
-                    '[header.%s] value must be a scalar (string|int|bool|float); got %s.',
-                    $name,
-                    get_debug_type($rawValue),
-                );
-
-                continue;
-            }
-
-            $coerced = TypeCoercer::coercePrimitive($rawValue, $schema);
-            $jsonSchema = OpenApiSchemaConverter::convert($schema, $version, SchemaContext::Request);
-
-            $schemaObject = ObjectConverter::convert($jsonSchema);
-            $dataObject = ObjectConverter::convert($coerced);
-
-            $formatted = $this->runner->validate($schemaObject, $dataObject);
-            foreach ($formatted as $path => $messages) {
-                $suffix = $path === '/' ? '' : $path;
-                foreach ($messages as $message) {
-                    $errors[] = "[header.{$name}{$suffix}] {$message}";
-                }
-            }
-        }
-
-        return $errors;
-    }
-
-    /**
-     * Validate the endpoint's `security` requirement against the incoming
-     * request. Supports `http` + `bearer` and `apiKey` (in: header|query|cookie)
-     * schemes. OAuth2 / OpenID Connect are out of scope for phase 1 and are
-     * treated as unsupported: any requirement entry containing an unsupported
-     * scheme is skipped entirely (contributes neither pass nor fail).
-     *
-     * Resolution: operation-level `security` takes precedence; otherwise the
-     * root-level `security` is inherited. `security: []` (empty array) explicitly
-     * opts out of all authentication, so it returns `[]` immediately regardless
-     * of root-level definitions. Missing `security` on both levels also returns
-     * `[]` (no authentication required).
-     *
-     * OR / AND semantics (per OpenAPI):
-     * - Multiple entries in the `security` array → OR (any one satisfied is enough)
-     * - Multiple schemes within a single entry object → AND (all must be satisfied)
-     *
-     * Malformed spec elements (undefined scheme references, scalar entries,
-     * missing `type` / `scheme` / `name` / `in` fields) are always surfaced as
-     * hard errors, even if another requirement entry is satisfied — a broken
-     * security declaration is something the spec author must fix.
-     *
-     * @param array<string, mixed> $spec full spec root (for `components.securitySchemes` + root-level `security`)
-     * @param array<string, mixed> $operation operation spec (for operation-level `security`)
-     * @param array<array-key, mixed> $headers caller-supplied request headers
-     * @param array<string, mixed> $queryParams parsed query string
-     * @param array<string, mixed> $cookies request cookies (for `apiKey` with `in: cookie`)
-     *
-     * @return string[]
-     */
-    private function validateSecurity(
-        string $method,
-        string $matchedPath,
-        array $spec,
-        array $operation,
-        array $headers,
-        array $queryParams,
-        array $cookies,
-    ): array {
-        $security = array_key_exists('security', $operation)
-            ? $operation['security']
-            : ($spec['security'] ?? null);
-
-        if ($security === null) {
-            return [];
-        }
-
-        if (!is_array($security)) {
-            return [
-                sprintf(
-                    '[security] %s %s: operation/root-level `security` must be an array of requirement objects, got %s.',
-                    $method,
-                    $matchedPath,
-                    get_debug_type($security),
-                ),
-            ];
-        }
-
-        if ($security === []) {
-            return [];
-        }
-
-        $schemes = $spec['components']['securitySchemes'] ?? [];
-        // A non-array `components.securitySchemes` is a malformed spec — without
-        // this hard error, every scheme reference below would be reported as
-        // "undefined scheme … add it under components.securitySchemes", which
-        // misdirects the spec author away from the real cause.
-        if (!is_array($schemes)) {
-            return [
-                sprintf(
-                    '[security] %s %s: components.securitySchemes must be an object mapping scheme names to definitions, got %s.',
-                    $method,
-                    $matchedPath,
-                    get_debug_type($schemes),
-                ),
-            ];
-        }
-
-        $normalizedHeaders = HeaderNormalizer::normalize($headers);
-
-        $hardErrors = [];
-        $failureErrors = [];
-        $satisfied = false;
-
-        foreach ($security as $entryIndex => $entry) {
-            if (!is_array($entry)) {
-                $hardErrors[] = sprintf(
-                    '[security] %s %s: security requirement at index %d must be an object mapping scheme names to scope arrays, got %s.',
-                    $method,
-                    $matchedPath,
-                    is_int($entryIndex) ? $entryIndex : 0,
-                    get_debug_type($entry),
-                );
-
-                continue;
-            }
-
-            $entryHasHardError = false;
-            $entryHasUnsupported = false;
-            /** @var array<string, array{kind: string, def: array<string, mixed>}> $validatable */
-            $validatable = [];
-
-            foreach ($entry as $schemeName => $_scopes) {
-                if (!is_string($schemeName)) {
-                    $hardErrors[] = sprintf(
-                        '[security] %s %s: security scheme name must be a string, got %s.',
-                        $method,
-                        $matchedPath,
-                        get_debug_type($schemeName),
-                    );
-                    $entryHasHardError = true;
-
-                    continue;
-                }
-
-                $schemeDef = $schemes[$schemeName] ?? null;
-                if (!is_array($schemeDef)) {
-                    $hardErrors[] = sprintf(
-                        "[security] %s %s: security requirement references undefined scheme '%s' — add it under components.securitySchemes.",
-                        $method,
-                        $matchedPath,
-                        $schemeName,
-                    );
-                    $entryHasHardError = true;
-
-                    continue;
-                }
-
-                $classification = $this->classifySecurityScheme($schemeDef);
-
-                if ($classification['kind'] === 'malformed') {
-                    $hardErrors[] = sprintf(
-                        "[security] %s %s: security scheme '%s' is malformed: %s",
-                        $method,
-                        $matchedPath,
-                        $schemeName,
-                        $classification['reason'],
-                    );
-                    $entryHasHardError = true;
-
-                    continue;
-                }
-
-                if ($classification['kind'] === 'unsupported') {
-                    $entryHasUnsupported = true;
-
-                    continue;
-                }
-
-                $validatable[$schemeName] = ['kind' => $classification['kind'], 'def' => $schemeDef];
-            }
-
-            if ($entryHasHardError) {
-                continue;
-            }
-
-            if ($entryHasUnsupported) {
-                continue;
-            }
-
-            $entryFailures = [];
-            foreach ($validatable as $schemeName => $info) {
-                $schemeErrors = $this->checkSchemeSatisfaction(
-                    $info['kind'],
-                    $info['def'],
-                    $normalizedHeaders,
-                    $queryParams,
-                    $cookies,
-                );
-                foreach ($schemeErrors as $schemeError) {
-                    $entryFailures[] = sprintf(
-                        "[security] %s %s: requirement '%s' not satisfied: %s",
-                        $method,
-                        $matchedPath,
-                        $schemeName,
-                        $schemeError,
-                    );
-                }
-            }
-
-            if ($entryFailures === []) {
-                $satisfied = true;
-
-                break;
-            }
-
-            $failureErrors = [...$failureErrors, ...$entryFailures];
-        }
-
-        if ($satisfied) {
-            return $hardErrors;
-        }
-
-        // If every requirement entry was either skipped (unsupported scheme
-        // within) or malformed (already captured in $hardErrors), there is no
-        // *validatable* entry that failed. Returning early avoids blocking a
-        // test for a spec we fundamentally cannot evaluate (false-negative
-        // avoidance for oauth2-only endpoints).
-        if ($failureErrors === []) {
-            return $hardErrors;
-        }
-
-        return [...$hardErrors, ...$failureErrors];
-    }
-
-    /**
-     * Classify a security scheme definition into one of:
-     * - `bearer`      — http + scheme=bearer (validatable)
-     * - `apiKey`      — apiKey in header|query|cookie (validatable)
-     * - `unsupported` — a spec-allowed type we intentionally defer (oauth2,
-     *                   openIdConnect, mutualTLS, or http with a non-bearer
-     *                   scheme). Phase 1 skip — false-negative avoidance.
-     * - `malformed`   — missing/invalid required fields, or a `type` that is
-     *                   not in the OpenAPI-enumerated set. A typo like
-     *                   `{"type": "htpp"}` MUST surface as a hard error
-     *                   rather than silently skipping — otherwise the
-     *                   library would pass every request for that endpoint.
-     *
-     * @param array<string, mixed> $schemeDef
-     *
-     * @return array{kind: string, reason?: string}
-     */
-    private function classifySecurityScheme(array $schemeDef): array
-    {
-        $type = $schemeDef['type'] ?? null;
-        if (!is_string($type) || $type === '') {
-            return ['kind' => 'malformed', 'reason' => "missing required 'type' field."];
-        }
-
-        if ($type === 'apiKey') {
-            $in = $schemeDef['in'] ?? null;
-            $name = $schemeDef['name'] ?? null;
-            if (!is_string($in) || !is_string($name)) {
-                return ['kind' => 'malformed', 'reason' => "apiKey scheme requires string 'in' and 'name' fields."];
-            }
-            if (!in_array($in, ['header', 'query', 'cookie'], true)) {
-                return ['kind' => 'malformed', 'reason' => "apiKey scheme 'in' must be one of header|query|cookie, got '{$in}'."];
-            }
-
-            return ['kind' => 'apiKey'];
-        }
-
-        if ($type === 'http') {
-            $scheme = $schemeDef['scheme'] ?? null;
-            if (!is_string($scheme)) {
-                return ['kind' => 'malformed', 'reason' => "http scheme requires a string 'scheme' field (e.g. 'bearer', 'basic')."];
-            }
-
-            if (strtolower($scheme) === 'bearer') {
-                return ['kind' => 'bearer'];
-            }
-
-            // http + basic / digest / etc. are well-formed but phase 1 cannot validate them.
-            return ['kind' => 'unsupported'];
-        }
-
-        if ($type === 'oauth2' || $type === 'openIdConnect' || $type === 'mutualTLS') {
-            return ['kind' => 'unsupported'];
-        }
-
-        return [
-            'kind' => 'malformed',
-            'reason' => "unknown type '{$type}' — OpenAPI 3.x enumerates apiKey|http|oauth2|openIdConnect|mutualTLS.",
-        ];
-    }
-
-    /**
-     * Check whether a single (already-classified, well-formed) security scheme
-     * is satisfied by the request. Returns an empty list when satisfied, or
-     * one or more error strings explaining why not.
-     *
-     * @param array<string, mixed> $schemeDef
-     * @param array<string, mixed> $normalizedHeaders lower-cased header map
-     * @param array<string, mixed> $queryParams
-     * @param array<string, mixed> $cookies
-     *
-     * @return string[]
-     */
-    private function checkSchemeSatisfaction(
-        string $kind,
-        array $schemeDef,
-        array $normalizedHeaders,
-        array $queryParams,
-        array $cookies,
-    ): array {
-        if ($kind === 'bearer') {
-            $raw = $normalizedHeaders['authorization'] ?? null;
-            $value = $this->extractSingleStringValue($raw);
-            if ($value === null || $value === '') {
-                return ['Authorization header is missing.'];
-            }
-
-            // RFC 6750 §2.1: `Bearer <token>`. Scheme name is case-insensitive
-            // per RFC 7235 §2.1, so we accept "Bearer" / "bearer" / "BEARER" etc.
-            // Require a non-empty token portion; "Bearer" alone or "Bearer " is
-            // not a valid credential.
-            if (preg_match('/^bearer\s+(\S+)/i', $value) !== 1) {
-                return ["Authorization header does not contain a 'Bearer <token>' credential."];
-            }
-
-            return [];
-        }
-
-        if ($kind === 'apiKey') {
-            /** @var string $in */
-            $in = $schemeDef['in'];
-            /** @var string $name */
-            $name = $schemeDef['name'];
-
-            $raw = match ($in) {
-                'header' => $normalizedHeaders[strtolower($name)] ?? null,
-                'query' => $queryParams[$name] ?? null,
-                'cookie' => $cookies[$name] ?? null,
-                default => null,
-            };
-
-            $value = $this->extractSingleStringValue($raw);
-            if ($value === null || $value === '') {
-                return [sprintf("api key '%s' is missing from the %s.", $name, $in)];
-            }
-
-            return [];
-        }
-
-        throw new LogicException("checkSchemeSatisfaction received unexpected kind '{$kind}'.");
-    }
-
-    /**
-     * Return the first element of an array, the value itself if it's a string,
-     * or `null` otherwise (absent, empty array, or non-string scalar like int
-     * or bool). No coercion is performed — a non-string first element still
-     * returns `null`.
-     *
-     * Unlike `validateHeaderParameters()` (which rejects multi-value arrays as
-     * a hard error to force the spec author to pick a canonical value), the
-     * security layer silently accepts the first element. Presence of a
-     * credential is all the security layer checks, and duplicate headers are
-     * a framework-level concern surfaced elsewhere.
-     */
-    private function extractSingleStringValue(mixed $value): ?string
-    {
-        if ($value === null) {
-            return null;
-        }
-
-        if (is_array($value)) {
-            if ($value === []) {
-                return null;
-            }
-
-            $first = $value[array_key_first($value)];
-
-            return is_string($first) ? $first : null;
-        }
-
-        return is_string($value) ? $value : null;
-    }
-
-    /**
-     * Merge path-level and operation-level parameters. Operation-level entries
-     * override path-level ones with the same `name` + `in` (per OpenAPI spec).
-     *
-     * Malformed entries are surfaced as errors rather than silently skipped,
-     * because for a contract-testing tool the absence of an error means
-     * "validated and OK" — silently dropping a parameter would leave drift
-     * invisible. `$ref` entries never reach this method: `OpenApiSpecLoader`
-     * resolves internal refs and throws on external/circular/unresolvable ones
-     * at load time.
-     *
-     * @param array<string, mixed> $pathSpec
-     * @param array<string, mixed> $operation
-     *
-     * @return array{0: list<array<string, mixed>>, 1: string[]}
-     */
-    private function collectParameters(string $method, string $matchedPath, array $pathSpec, array $operation): array
-    {
-        $merged = [];
-        $errors = [];
-
-        foreach ([$pathSpec['parameters'] ?? [], $operation['parameters'] ?? []] as $source) {
-            if (!is_array($source)) {
-                continue;
-            }
-
-            foreach ($source as $param) {
-                if (!is_array($param)) {
-                    $errors[] = "Malformed parameter entry for {$method} {$matchedPath}: expected object, got scalar.";
-
-                    continue;
-                }
-
-                if (!isset($param['in'], $param['name']) || !is_string($param['in']) || !is_string($param['name'])) {
-                    $errors[] = "Malformed parameter entry for {$method} {$matchedPath}: 'name' and 'in' must be strings.";
-
-                    continue;
-                }
-
-                // OAS 3.x §4.7.12.1 says `Accept`/`Content-Type`/`Authorization`
-                // in:header parameter definitions SHALL be ignored — content
-                // negotiation and security schemes own them. Silently skipping
-                // would let a spec author's `required: true` pass every request;
-                // surface as a hard spec error and drop the entry so downstream
-                // per-request validation stays OAS-compliant (no runtime effect).
-                // The `in === 'header'` guard keeps a non-header parameter that
-                // coincidentally shares a reserved name (e.g. `in: query, name: accept`)
-                // out of scope — only in:header entries are governed by §4.7.12.1.
-                if ($param['in'] === 'header' && in_array(strtolower($param['name']), self::RESERVED_HEADER_NAMES, true)) {
-                    $errors[] = sprintf(
-                        'Reserved in:header parameter declared for %s %s: [header.%s] — per OpenAPI 3.x §4.7.12.1, `Accept`/`Content-Type`/`Authorization` parameter definitions SHALL be ignored. Remove the declaration or use `content` / security schemes instead.',
-                        $method,
-                        $matchedPath,
-                        $param['name'],
-                    );
-
-                    continue;
-                }
-
-                $key = $param['in'] . ':' . $param['name'];
-                $merged[$key] = $param;
-            }
-        }
-
-        return [array_values($merged), $errors];
-    }
-
-    /**
-     * Validate the request body against the operation's `requestBody` schema.
-     *
-     * Returns an empty list when the body is acceptable (including when the
-     * spec defines no body, no content, no JSON content type, or no schema).
-     * Hard spec-level errors (malformed `requestBody` / `content`) are
-     * reported as standard error entries so they compose with query errors.
-     *
-     * @param array<string, mixed> $operation
-     *
-     * @return string[]
-     */
-    private function validateRequestBody(
-        string $specName,
-        string $method,
-        string $matchedPath,
-        array $operation,
-        mixed $requestBody,
-        ?string $contentType,
-        OpenApiVersion $version,
-    ): array {
-        // OpenAPI: a missing requestBody means the operation accepts no body — treat as success.
-        if (!isset($operation['requestBody'])) {
-            return [];
-        }
-
-        // A present-but-non-array requestBody signals a malformed spec (stray scalar).
-        // Contract-testing tools should surface this, not mask it as "no body".
-        if (!is_array($operation['requestBody'])) {
-            return [
-                "Malformed 'requestBody' for {$method} {$matchedPath} in '{$specName}' spec: expected object, got scalar.",
-            ];
-        }
-
-        /** @var array<string, mixed> $requestBodySpec */
-        $requestBodySpec = $operation['requestBody'];
-
-        $required = ($requestBodySpec['required'] ?? false) === true;
-
-        if (!isset($requestBodySpec['content'])) {
-            return [];
-        }
-
-        if (!is_array($requestBodySpec['content'])) {
-            return [
-                "Malformed 'requestBody.content' for {$method} {$matchedPath} in '{$specName}' spec: expected object, got scalar.",
-            ];
-        }
-
-        /** @var array<string, mixed> $content */
-        $content = $requestBodySpec['content'];
-
-        foreach ($content as $mediaType => $mediaTypeSpec) {
-            // The @var on $content narrows values to array, but PHPDoc is unchecked at
-            // runtime — a malformed spec like `content: {"application/json": "oops"}`
-            // would TypeError on downstream array accesses. Surface it as a loud spec
-            // error instead, matching the sibling guard on `requestBody.content` above.
-            if (!is_array($mediaTypeSpec)) {
-                return [
-                    "Malformed 'requestBody.content[\"{$mediaType}\"]' for {$method} {$matchedPath} in '{$specName}' spec: expected object, got scalar.",
-                ];
-            }
-
-            // `schema: "oops"` (or any other non-array scalar) would slip past the
-            // downstream `isset(...['schema'])` presence check and reach
-            // OpenApiSchemaConverter::convert() as a scalar, producing a confusing
-            // TypeError instead of a spec-level error. array_key_exists rather than
-            // isset so an explicit `schema: null` is also flagged.
-            if (array_key_exists('schema', $mediaTypeSpec) && !is_array($mediaTypeSpec['schema'])) {
-                return [
-                    "Malformed 'requestBody.content[\"{$mediaType}\"].schema' for {$method} {$matchedPath} in '{$specName}' spec: expected object, got scalar.",
-                ];
-            }
-        }
-
-        // When the actual request Content-Type is provided, handle content negotiation:
-        // non-JSON types are checked for spec presence only, while JSON-compatible types
-        // fall through to schema validation against the first JSON media type in the spec.
-        if ($contentType !== null) {
-            $normalizedType = ContentTypeMatcher::normalizeMediaType($contentType);
-
-            if (!ContentTypeMatcher::isJsonContentType($normalizedType)) {
-                if (ContentTypeMatcher::isContentTypeInSpec($normalizedType, $content)) {
-                    return [];
-                }
-
-                $defined = implode(', ', array_keys($content));
-
-                return [
-                    "Request Content-Type '{$normalizedType}' is not defined for {$method} {$matchedPath} in '{$specName}' spec. Defined content types: {$defined}",
-                ];
-            }
-
-            // JSON-compatible request: fall through to existing JSON schema validation.
-            // JSON types are treated as interchangeable (e.g. application/vnd.api+json
-            // validates against an application/json spec entry) because the schema is
-            // the same regardless of the specific JSON media type.
-        }
-
-        $jsonContentType = ContentTypeMatcher::findJsonContentType($content);
-
-        // If no JSON-compatible content type is defined, skip body validation.
-        // This validator only handles JSON schemas; non-JSON types (e.g. application/xml,
-        // application/octet-stream) are outside its scope.
-        if ($jsonContentType === null) {
-            return [];
-        }
-
-        if (!isset($content[$jsonContentType]['schema'])) {
-            return [];
-        }
-
-        if ($requestBody === null) {
-            if (!$required) {
-                return [];
-            }
-
-            return [
-                "Request body is empty but {$method} {$matchedPath} defines a required JSON request body schema in '{$specName}' spec.",
-            ];
-        }
-
-        /** @var array<string, mixed> $schema */
-        $schema = $content[$jsonContentType]['schema'];
-        $jsonSchema = OpenApiSchemaConverter::convert($schema, $version, SchemaContext::Request);
-
-        $schemaObject = ObjectConverter::convert($jsonSchema);
-        $dataObject = ObjectConverter::convert($requestBody);
-
-        $formatted = $this->runner->validate($schemaObject, $dataObject);
-
-        $errors = [];
-        foreach ($formatted as $path => $messages) {
-            foreach ($messages as $message) {
-                $errors[] = "[{$path}] {$message}";
-            }
-        }
-
-        return $errors;
     }
 }
