+++ tests/Unit/Validation/Request/HeaderParameterValidatorTest.php
@@ -1,0 +1,78 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Studio\OpenApiContractTesting\Tests\Unit\Validation\Request;
+
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\TestCase;
+use Studio\OpenApiContractTesting\OpenApiVersion;
+use Studio\OpenApiContractTesting\Validation\Request\HeaderParameterValidator;
+use Studio\OpenApiContractTesting\Validation\Support\SchemaValidatorRunner;
+
+class HeaderParameterValidatorTest extends TestCase
+{
+    private HeaderParameterValidator $validator;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->validator = new HeaderParameterValidator(new SchemaValidatorRunner(20));
+    }
+
+    #[Test]
+    public function validate_matches_headers_case_insensitively(): void
+    {
+        $parameters = [
+            ['name' => 'X-Request-Id', 'in' => 'header', 'required' => true, 'schema' => ['type' => 'string']],
+        ];
+
+        $errors = $this->validator->validate(
+            'GET',
+            '/pets',
+            $parameters,
+            ['x-request-id' => 'abc'],
+            OpenApiVersion::V3_0,
+        );
+
+        $this->assertSame([], $errors);
+    }
+
+    #[Test]
+    public function validate_rejects_multi_value_header_with_scalar_schema(): void
+    {
+        $parameters = [
+            ['name' => 'X-Foo', 'in' => 'header', 'schema' => ['type' => 'string']],
+        ];
+
+        $errors = $this->validator->validate(
+            'GET',
+            '/pets',
+            $parameters,
+            ['X-Foo' => ['a', 'b']],
+            OpenApiVersion::V3_0,
+        );
+
+        $this->assertCount(1, $errors);
+        $this->assertStringContainsString('multiple values received', $errors[0]);
+    }
+
+    #[Test]
+    public function validate_flags_non_scalar_header_value(): void
+    {
+        $parameters = [
+            ['name' => 'X-Foo', 'in' => 'header', 'schema' => ['type' => 'string']],
+        ];
+
+        $errors = $this->validator->validate(
+            'GET',
+            '/pets',
+            $parameters,
+            ['X-Foo' => [['nested']]],
+            OpenApiVersion::V3_0,
+        );
+
+        $this->assertCount(1, $errors);
+        $this->assertStringContainsString('value must be a scalar', $errors[0]);
+    }
+}
